The game.cpp file now runs

+++ Assignments/10-P03/diceRoll.hpp
@@ -44,7 +44,7 @@ public:
         return true;
     }
 
-    size_t getRandomFrame()
+    int getRandomFrame()
     {
         return randomFrame;
     }
